refactor add global response status

+++ respones/user_response.go
@@ -1,7 +1,7 @@
 package respones
 
 type CustomResponse struct {
-	Status      int         `json:"status"`
+	Status      int16       `json:"status"`
 	Message     string      `json:"message"`
 	Description string      `json:"description"`
 	Data        interface{} `json:"data"`
